Remove server.js
